i64 f64 now becomes i32 and f32. Implementing BigInts later

+++ src/eval.rs
@@ -122,9 +122,9 @@ pub fn exec_binary_op(
     };
 }
 
-fn return_correct_lnum_type(res: f64) -> LRes<Obj> {
+fn return_correct_lnum_type(res: f32) -> LRes<Obj> {
     if res.fract() == 0.0 {
-        Ok(Obj::Num(LNum::Int(res as i64)))
+        Ok(Obj::Num(LNum::Int(res as i32)))
     } else {
         Ok(Obj::Num(LNum::Float(res)))
     }
